+++ src/frontend/src/optimizer/plan_node/mod.rs
@@ -91,6 +91,10 @@ impl GenericPlanRef for PlanRef {
     fn append_only(&self) -> bool {
         self.plan_base().append_only
     }
+
+    fn logical_pk(&self) -> &[usize] {
+        &self.plan_base().logical_pk
+    }
 }
 
 impl dyn PlanNode {
